[fix](cloud) modify primary cluster bes to be in CloudReplica to reduce memory (#59932)
